move operations related to brillig-ssa conversion to brillig_gen
