+++ crates/noirc_evaluator/src/brillig/brillig_ir.rs
@@ -5,5 +5,12 @@
 //! ssa types and types in this module.
 //! A similar paradigm can be seen with the `acir_ir` module.
 pub(crate) mod artifact;
-pub(crate) mod binary;
 pub(crate) mod memory;
+
+use acvm::acir::brillig_vm::{BinaryFieldOp, BinaryIntOp};
+
+/// Type to encapsulate the binary operation types in Brillig
+pub(crate) enum BrilligBinaryOp {
+    Field { op: BinaryFieldOp },
+    Integer { op: BinaryIntOp, bit_size: u32 },
+}
